requirements.txt
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
lxml==4.9.0
python-decouple==3.6
random-user-agent==1.0.1
requests==2.28.0
soupsieve==2.3.2.post1
tqdm==4.64.0
urllib3==1.26.9
